Add global exception handling (JobNotFoundException, GlobalExceptionHandler)

+++ src/main/java/com/ayhanekin/jobapi/service/JobService.java
@@ -3,10 +3,12 @@ package com.ayhanekin.jobapi.service;
 import com.ayhanekin.jobapi.dto.request.CreateJobRequest;
 import com.ayhanekin.jobapi.dto.request.UpdateJobRequest;
 import com.ayhanekin.jobapi.dto.response.JobResponse;
+import com.ayhanekin.jobapi.exception.JobNotFoundException;
 import com.ayhanekin.jobapi.mapper.JobMapper;
 import com.ayhanekin.jobapi.model.JobPost;
 import com.ayhanekin.jobapi.repo.JobRepo;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.UUID;
@@ -24,24 +26,30 @@ public class JobService {
     }
 
     public JobResponse getJobPost(UUID id) {
-        return mapper.toResponse(repo.findById(id).orElse(new JobPost()));
+        return mapper.toResponse(repo.findById(id)
+                .orElseThrow(() -> new JobNotFoundException(id)));
     }
 
     public JobResponse save(CreateJobRequest request) {
         return mapper.toResponse(repo.save(mapper.createJobPost(request)));
     }
 
-    public void update(UUID id, UpdateJobRequest jobRequest) {
+    public JobResponse update(UUID id, UpdateJobRequest jobRequest) {
         JobPost jobPost = repo.findById(id).orElseThrow(() ->
-                (new RuntimeException("Job not found")));
-
+                (new JobNotFoundException(id)));
         mapper.updateJobPost(jobRequest,jobPost);
         repo.save(jobPost);
 
 
+        return mapper.toResponse(jobPost);
+
     }
 
     public void delete(UUID id) {
-        repo.deleteById(id);
+        if (!repo.existsById(id)){
+            throw new JobNotFoundException(id);
+        }else {
+            repo.deleteById(id);
+        }
     }
 }
